add some for markdown and re-use <up>&<down>
There are more features for markdown usage now. And I supply some plugs for markdown additionally.
I reuse the <up> & <down> mappings to resize the windows. It is no longer throwing bugs on me.
In this commit, if you compile CPP files, u may be surprised!
You can see how to use fuzzy file finder at the beginning now.
VIM 8.1 users may use tr to open a terminal in vim, although it is not that one you are using.
Take a glance from Line 194 to 203 and 268.